+++ backend/src/prisma/migrations/20260904031524_add_published_at_column_on_events_table/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE `events` ADD COLUMN `published_at` DATETIME(3) NULL;